+++ lib/rest/conversations/v1/conversation/message.d.ts
@@ -184,6 +184,8 @@ interface MessageListInstance {
  * @property attributes - A string metadata field you can use to store any data you wish.
  * @property author - The channel specific identifier of the message's author.
  * @property body - The content of the message.
+ * @property contentSid - The unique ID of the multi-channel Rich Content template.
+ * @property contentVariables - A structurally valid JSON string that contains values to resolve Rich Content template variables.
  * @property dateCreated - The date that this resource was created.
  * @property dateUpdated - The date that this resource was last updated.
  * @property mediaSid - The Media SID to be attached to the new Message.
@@ -193,6 +195,8 @@ interface MessageListInstanceCreateOptions {
   attributes?: string;
   author?: string;
   body?: string;
+  contentSid?: string;
+  contentVariables?: string;
   dateCreated?: Date;
   dateUpdated?: Date;
   mediaSid?: string;
@@ -270,6 +274,7 @@ interface MessageResource {
   attributes: string;
   author: string;
   body: string;
+  content_sid: string;
   conversation_sid: string;
   date_created: Date;
   date_updated: Date;
@@ -353,6 +358,7 @@ declare class MessageInstance extends SerializableClass {
   attributes: string;
   author: string;
   body: string;
+  contentSid: string;
   conversationSid: string;
   dateCreated: Date;
   dateUpdated: Date;