+++ spec/integration/rest/conversations/v1/conversation/message.spec.js
@@ -76,6 +76,7 @@ describe('Message', function() {
               'failed': 'none',
               'undelivered': 'none'
           },
+          'content_sid': null,
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -123,6 +124,7 @@ describe('Message', function() {
               'failed': 'none',
               'undelivered': 'none'
           },
+          'content_sid': null,
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -163,6 +165,48 @@ describe('Message', function() {
               'failed': 'none',
               'undelivered': 'none'
           },
+          'content_sid': null,
+          'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'links': {
+              'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
+          }
+      };
+
+      holodeck.mock(new Response(201, body));
+
+      var promise = client.conversations.v1.conversations('CHXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                           .messages.create();
+      promise.then(function(response) {
+        expect(response).toBeDefined();
+        done();
+      }, function() {
+        throw new Error('failed');
+      }).done();
+    }
+  );
+  it('should generate valid create_with_content_sid response',
+    function(done) {
+      var body = {
+          'sid': 'IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'conversation_sid': 'CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'body': 'Hello John',
+          'media': null,
+          'author': 'message author',
+          'participant_sid': 'MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'attributes': '{}',
+          'date_created': '2015-12-16T22:18:37Z',
+          'date_updated': '2015-12-16T22:18:38Z',
+          'index': 0,
+          'delivery': {
+              'total': 2,
+              'sent': 'all',
+              'delivered': 'some',
+              'read': 'some',
+              'failed': 'none',
+              'undelivered': 'none'
+          },
+          'content_sid': 'HXaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -229,6 +273,7 @@ describe('Message', function() {
               'failed': 'none',
               'undelivered': 'none'
           },
+          'content_sid': null,
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -334,6 +379,7 @@ describe('Message', function() {
               'failed': 'none',
               'undelivered': 'none'
           },
+          'content_sid': null,
           'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -385,6 +431,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -410,6 +457,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -442,6 +490,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -487,6 +536,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -512,6 +562,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -544,6 +595,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -594,6 +646,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -619,6 +672,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -651,6 +705,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -718,6 +773,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -743,6 +799,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -775,6 +832,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'
@@ -835,6 +893,7 @@ describe('Message', function() {
                       'failed': 'none',
                       'undelivered': 'none'
                   },
+                  'content_sid': null,
                   'url': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'delivery_receipts': 'https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts'