+++ spec/integration/rest/messaging/v1/tollfreeVerification.spec.js
@@ -76,7 +76,7 @@ describe('TollfreeVerification', function() {
           'business_contact_phone': '+16504988765',
           'notification_email': 'vikram@gmail.com',
           'use_case_categories': [
-              '2FA',
+              'TWO_FACTOR_AUTHENTICATION',
               'MARKETING'
           ],
           'use_case_summary': 'test',
@@ -144,7 +144,7 @@ describe('TollfreeVerification', function() {
                   'business_contact_phone': '+16504988765',
                   'notification_email': 'vikram@gmail.com',
                   'use_case_categories': [
-                      '2FA',
+                      'TWO_FACTOR_AUTHENTICATION',
                       'MARKETING'
                   ],
                   'use_case_summary': 'test',
@@ -206,7 +206,7 @@ describe('TollfreeVerification', function() {
                   'business_contact_phone': '+16504988765',
                   'notification_email': 'vikram@gmail.com',
                   'use_case_categories': [
-                      '2FA',
+                      'TWO_FACTOR_AUTHENTICATION',
                       'MARKETING'
                   ],
                   'use_case_summary': 'test',
@@ -273,7 +273,7 @@ describe('TollfreeVerification', function() {
                   'business_contact_phone': '+16504988765',
                   'notification_email': 'vikram@gmail.com',
                   'use_case_categories': [
-                      '2FA',
+                      'TWO_FACTOR_AUTHENTICATION',
                       'MARKETING'
                   ],
                   'use_case_summary': 'test',
@@ -355,7 +355,7 @@ describe('TollfreeVerification', function() {
                   'business_contact_phone': '+16504988765',
                   'notification_email': 'vikram@gmail.com',
                   'use_case_categories': [
-                      '2FA',
+                      'TWO_FACTOR_AUTHENTICATION',
                       'MARKETING'
                   ],
                   'use_case_summary': 'test',
@@ -460,7 +460,7 @@ describe('TollfreeVerification', function() {
           'business_contact_phone': '+16504988765',
           'notification_email': 'vikram@gmail.com',
           'use_case_categories': [
-              '2FA',
+              'TWO_FACTOR_AUTHENTICATION',
               'MARKETING'
           ],
           'use_case_summary': 'test',
@@ -492,6 +492,80 @@ describe('TollfreeVerification', function() {
         'tollfreePhoneNumberSid': 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX'
       };
       var promise = client.messaging.v1.tollfreeVerifications.create(opts);
+      promise.then(function(response) {
+        expect(response).toBeDefined();
+        done();
+      }, function() {
+        throw new Error('failed');
+      }).done();
+    }
+  );
+  it('should generate valid update request',
+    function(done) {
+      holodeck.mock(new Response(500, {}));
+
+      var promise = client.messaging.v1.tollfreeVerifications('HHXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').update();
+      promise.then(function() {
+        throw new Error('failed');
+      }, function(error) {
+        expect(error.constructor).toBe(RestException.prototype.constructor);
+        done();
+      }).done();
+
+      var sid = 'HHXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX';
+      var url = `https://messaging.twilio.com/v1/Tollfree/Verifications/${sid}`;
+
+      holodeck.assertHasRequest(new Request({
+        method: 'POST',
+        url: url
+      }));
+    }
+  );
+  it('should generate valid update response',
+    function(done) {
+      var body = {
+          'sid': 'HHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'regulated_item_sid': null,
+          'customer_profile_sid': 'BU3344409f7e067e279523808d267e2d85',
+          'trust_product_sid': null,
+          'status': 'PENDING_REVIEW',
+          'date_created': '2021-01-27T14:18:35Z',
+          'date_updated': '2021-01-27T14:18:36Z',
+          'business_name': 'Agent',
+          'business_street_address': '927 Terrace St',
+          'business_street_address2': 'Unit 4',
+          'business_city': 'Tempe',
+          'business_state_province_region': 'AZ',
+          'business_postal_code': '85281',
+          'business_country': 'USA',
+          'business_website': 'www.ghost.com',
+          'business_contact_first_name': 'Vikram',
+          'business_contact_last_name': 'Amar',
+          'business_contact_email': 'vikram@gmail.com',
+          'business_contact_phone': '+16504988765',
+          'notification_email': 'vikram@gmail.com',
+          'use_case_categories': [
+              'TWO_FACTOR_AUTHENTICATION',
+              'MARKETING'
+          ],
+          'use_case_summary': 'test',
+          'production_message_sample': 'test1',
+          'opt_in_image_urls': [
+              'https://zipwhiptestbusiness.com/images/image1.jpg',
+              'https://zipwhiptestbusiness.com/images/image2.jpg'
+          ],
+          'opt_in_type': 'VERBAL',
+          'message_volume': '1,000',
+          'additional_information': 'info',
+          'tollfree_phone_number_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'resource_links': {},
+          'url': 'https://messaging.twilio.com/v1/Tollfree/Verifications/HHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+      };
+
+      holodeck.mock(new Response(200, body));
+
+      var promise = client.messaging.v1.tollfreeVerifications('HHXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').update();
       promise.then(function(response) {
         expect(response).toBeDefined();
         done();